Cruds added to new handin2 dab

+++ DAB/Handin2/Handin2_new/Personkartotek/Personkartotek/DomainModel/AA.cs
@@ -10,7 +10,7 @@ namespace Personkartotek.DomainModel
     {
         public virtual string addressType { get; set; }
 
-        public virtual int addressId { get; set; }
+        public virtual long addressId { get; set; }
 
         public virtual List<Person> attachedPerson { get; set; }
